melhorias no core do server
