+++ .gitignore
@@ -107,7 +107,7 @@ ipython_config.py
 # Pipfile.lock
 
 # poetry
-poetry.lock
+# poetry.lock
 
 # PEP 582; used by e.g. github.com/David-OConnor/pyflow
 __pypackages__/
