+++ api_server/core/config/base.py
@@ -1,10 +1,16 @@
 import secrets
 from typing import Any, Dict, List, Optional
 
-from pydantic import PostgresDsn, BaseSettings, validator
+from pydantic import PostgresDsn
+from pydantic import BaseSettings as PydanticBaseSettings
+from pydantic import validator
 
 
-class Settings(BaseSettings):
+class BaseSettings(PydanticBaseSettings):
+
+    SERVER_TITLE: str = 'API Manager'
+    SERVER_VERSION: str = '1.0.0'
+    SERVER_DESCRIPTION: str = 'OpenAPI schema'
 
     API_PREFIX: str = '/api'
     SECRET_KEY: str = secrets.token_urlsafe(32)
@@ -13,6 +19,7 @@ class Settings(BaseSettings):
     ACCESS_TOKEN_EXPIRE_MINUTES: int = 60 * 24 * 8
 
     POSTGRES_SERVER: str
+    POSTGRES_PORT: str = '5432'
     POSTGRES_USER: str
     POSTGRES_PASSWORD: str
     POSTGRES_DB: str
@@ -29,6 +36,7 @@ class Settings(BaseSettings):
             user=values.get('POSTGRES_USER'),
             password=values.get('POSTGRES_PASSWORD'),
             host=values.get('POSTGRES_SERVER'),
+            port=values.get('POSTGRES_PORT'),
             path=f"/{values.get('POSTGRES_DB') or ''}",
         )
 
@@ -39,11 +47,6 @@ class Settings(BaseSettings):
 
     class Config:
         case_sensitive = True
-        env_file = '../local.env'
-        env_file_encoding = 'utf-8'
 
 
-settings = Settings()
-
-
-__all__ = ('settings',)
+__all__ = ('BaseSettings',)
