+++ api_server/core/config/production.py
@@ -1,0 +1,9 @@
+from core.config.base import BaseSettings
+
+
+class Settings(BaseSettings):
+
+    DEBUG: bool = False
+
+
+__all__ = ('Settings',)
