+++ api_server/debug.py
@@ -1,4 +1,4 @@
 import uvicorn
 
 if __name__ == '__main__':
-    uvicorn.run('main:app', host='0.0.0.0', port=5000, reload=True)
+    uvicorn.run('main:create_app', host='0.0.0.0', port=5000, reload=True, factory=True)
