+++ api_server/main.py
@@ -1,3 +1,36 @@
-from core.server import app
+from fastapi import FastAPI
 
-__all__ = ('app',)
+
+def create_app(config: str = 'development') -> FastAPI:
+
+    from core.config import create_config
+
+    settings = create_config(config)
+
+    from core import openapi, routers, handlers
+    from core.databases.sqlalchemy import database
+
+    app = FastAPI(
+        openapi_url=f'{settings.API_PREFIX}/openapi.json',
+        # docs_url=None,
+        # redoc_url=f'{settings.API_PREFIX}/documentation',
+        docs_url=f'{settings.API_PREFIX}/docs',
+        redoc_url=f'{settings.API_PREFIX}/redoc',
+    )
+
+    @app.on_event('startup')
+    async def startup():
+        await database.connect()
+
+    @app.on_event('shutdown')
+    async def shutdown():
+        await database.disconnect()
+
+    handlers.init_app(app)
+    routers.init_app(app)
+    openapi.init_app(app)
+
+    return app
+
+
+__all__ = ('create_app',)
